renamed logging directory, bug with def main() in init files fixed.
