+++ pyproject.toml
@@ -76,7 +76,7 @@ line-length = 96
 indent-width = 4
 [tool.ruff.lint]
 select = ["E", "F", "B", "I", "N", "UP", "YTT", "ANN", "S", "A", "D"]
-ignore = ["ANN102", "D100", "D101"]
+ignore = ["ANN102", "D100", "D101", "D104"]
 
 [tool.ruff.format]
 quote-style = "double"
