--- src/__init__.py
@@ -1,5 +1,0 @@
-def main() -> None:
-    ...
-
-if __name__ == '__main__':
-    main()
